+++ src/main/java/me/poma123/globalwarming/api/biomes/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains classes related to {@link io.github.thebusybiscuit.slimefun4.utils.biomes.BiomeMap}
+ */
+package me.poma123.globalwarming.api.biomes;
